requirements.txt
accelerate
accelerator
animus @ git+https://github.com/Scitator/animus@7c80dba79d9cac87455f5bc61c8bd4110dfb26f8
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
Babel
beautifulsoup4
bleach
bottle
certifi
cffi
charset-normalizer
coco-assistant
comet-ml
comm
configobj
contourpy
cycler
Cython
debugpy
decorator
defusedxml
dulwich
efficientnet-pytorch
everett
executing
fastjsonschema
filelock
fonttools
fqdn
fsspec
huggingface-hub
idna
imageio
iniconfig
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
kiwisolver
lazy_loader
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
mpmath
munch
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
overrides
packaging
pandas
pandocfilters
parso
pexpect
Pillow
platformdirs
pluggy
pretrainedmodels
prometheus-client
prompt-toolkit
psutil
ptyprocess
pure-eval
pycocotools
pycparser
PyExifTool
Pygments
pyparsing
pytest
python-box
python-dateutil
python-dotenv
python-json-logger
pytz
PyYAML
pyzmq
qtconsole
QtPy
referencing
requests
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rpds-py
safetensors
scikit-image
scipy
seaborn
segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@6db76a1106426ac5b55f39fba68168f3bccae7f8
semantic-version
Send2Trash
sentry-sdk
setproctitle
simplejson
six
sniffio
soupsieve
stack-data
sympy
terminado
tifffile
timm
tinycss2
torch
torchaudio
torchvision
tornado
tqdm
traitlets
triton
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
waitress
wcwidth
webcolors
webencodings
websocket-client
widgetsnbextension
wrapt
wurlitzer
